cxcmpl

+++ scripts/crosscompile/Host_syncwithTarget.py
@@ -6,8 +6,12 @@ import shutil
 
 # do not run with sudo
 
+sysroot_fn = "sysroot-relativelinks.py"
+
 def copy_if_not_exists(source, destination):
     if not os.path.exists(destination):
+        if not os.path.exists(source):
+            raise Exception(f"Source file {source} does not exist")
         shutil.copyfile(source, destination)
         print(f"Copied {source} to {destination}")
